+++ packages/stream_chat_localizations/example/lib/add_new_lang.dart
@@ -638,6 +638,20 @@ class NnStreamChatLocalizations extends GlobalStreamChatLocalizations {
 
   @override
   String get holdToRecordLabel => 'Hold to record, release to send.';
+
+  @override
+  String get sendAnywayLabel => 'Send Anyway';
+
+  @override
+  String get moderatedMessageBlockedText =>
+      'Message was blocked by moderation policies';
+
+  @override
+  String get moderationReviewModalTitle => 'Are you sure?';
+
+  @override
+  String get moderationReviewModalDescription =>
+      'Consider how your comment might make others feel and be sure to follow our Community Guidelines.';
 }
 
 void main() async {